Save camera image as a file
**Still can't test this stuff**
The reason this is important (aside from being a feature we might want
anyway) is because the way I save the image here saves a larger version
of the image than the one in the last commit. This means that when we
apply halftoning to the image, we should use THIS version, not the one I
display in last commit's code.