New test and small layout changes
I made a copy of the sample test that puts the app in landscape mode
first. That made me realise that the app could be laid out better. So I
did that. Later on we should make it so the layout changes depending on
whether the app is in portrait or landscape mode.