Layout looks great! Also, a readme
We need a readme with basic information as part of our submission, so
that's done (except for known bugs/limitations, which I don't feel we
really know yet). Also, I made the layout look great again by adding in
a scrollview to the output image - now it isn't absurdly tiny on smaller
devices/in landscape mode!